set MCP4822 dac working and basic shift indicator

+++ cubeide_proj/ugfx/drivers/gdisp/s6e63d6/gdisp_lld_s6e63d6.c
@@ -384,7 +384,7 @@ LLDSPEC void gdisp_lld_clear (GDisplay *g)
   }
 
   uint16_t c = map_color(gdispColor2Native(g->p.color));
-  uint32_t c2 = c | c<<16;
+  uint32_t c2 = (uint32_t)c | ((uint32_t)c)<<16;
   setClearColor(c2);
 
   // dont actually clear, it happened automatically during flushing
